fixed warnings: when compiling without openGL / when compiling with gcc-3.4
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@7630 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/console/OSDImageBasedWidget.cc
@@ -114,8 +114,8 @@ void OSDImageBasedWidget::transformXY(const OutputSurface& output,
 		width  = 0;
 		height = 0;
 	}
-	outx = x + getX() + relx * width;
-	outy = y + getY() + rely * height;
+	outx = x + getX() + int(relx * width);
+	outy = y + getY() + int(rely * height);
 	getParent()->transformXY(output, outx, outy, getRelX(), getRelY(),
 	                         outx, outy);
 }
